requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
pypdf
unstructured 
sentence-transformers
pdfminer.six
pi-heif